fix solaris problem -> 0.2.7

+++ source/src/Forest/ForestClassification.cpp
@@ -71,8 +71,8 @@ void ForestClassification::initInternal(std::string status_variable_name) {
 
   // If mtry not set, use floored square root of number of independent variables.
   if (mtry == 0) {
-    int temp = sqrt(num_variables - 1);
-    mtry = std::max(1, temp);
+    unsigned long temp = sqrt((double) (num_variables - 1));
+    mtry = std::max((unsigned long) 1, temp);
   }
 
   // Set minimal node size
