+++ source/src/Forest/ForestProbability.cpp
@@ -65,8 +65,8 @@ void ForestProbability::initInternal(std::string status_variable_name) {
 
   // If mtry not set, use floored square root of number of independent variables.
   if (mtry == 0) {
-    int temp = sqrt(num_variables - 1);
-    mtry = std::max(1, temp);
+    unsigned long temp = sqrt((double) (num_variables - 1));
+    mtry = std::max((unsigned long) 1, temp);
   }
 
   // Set minimal node size
