+++ source/src/Forest/ForestSurvival.cpp
@@ -78,8 +78,8 @@ void ForestSurvival::initInternal(std::string status_variable_name) {
 
   // If mtry not set, use floored square root of number of independent variables.
   if (mtry == 0) {
-    int temp = ceil(sqrt(num_variables - 2));
-    mtry = std::max(1, temp);
+    unsigned long temp = ceil(sqrt((double) (num_variables - 2)));
+    mtry = std::max((unsigned long) 1, temp);
   }
 
   // Set minimal node size
